database updated long cpr

+++ RESTfull/Models/Step.cs
@@ -17,7 +17,7 @@ namespace RESTfull.Models
         public int id { get; set; }
         public System.DateTime date { get; set; }
         public int count { get; set; }
-        public int clientId { get; set; }
+        public long clientId { get; set; }
     
         public virtual Client Client { get; set; }
     }
